Final Commit

+++ controller/employee.js
@@ -162,6 +162,9 @@ async function GenerateReports(req, res, next) {
 async function rem_employee(req, res, next) {
     try{
         console.log(req.body.usremail);
+        emp = EMP.findOne(req.body.usremail);
+        if(!emp)
+            return res.send(`No email found for ${req.body.usremail}`);
         await EMP.deleteOne({email: req.body.usremail})
         return res.send("Deleted");
     }
